+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [unreleased]
+
+### Added
+
+- Added a window to accept the EULA.
+- Added a non-installer zip for Windows.
+- 
+### Changed
+
+- Distribute the app inside a DMG for macOS.
+
 ## [v0.18.0] - November 1, 2025
 
 ### Added